fix: SecurityConfig reissue 허용경로 수정

See merge request s11-final/S11P31E204!227

+++ back/src/main/java/com/ssafy/daily/reward/entity/Shell.java
@@ -22,7 +22,7 @@ public class Shell {
     private Member member;
 
     @Column(nullable = false)
-    private byte stock;
+    private int stock;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
